Move BaseElement and UIElement to external folder
