+++ packages/lib/src/components/external/BaseElement.ts
@@ -1,8 +1,8 @@
 import { ComponentChild, render } from 'preact';
-import EventEmitter from '../../utils/EventEmitter';
-import uuid from '../../../utils/uuid';
-import Core from '../../../core';
-import { BaseElementProps, BaseElementState } from '../../types';
+import EventEmitter from './EventEmitter';
+import uuid from '../../utils/uuid';
+import Core from '../../core';
+import { BaseElementProps, BaseElementState } from '../types';
 
 class BaseElement<P extends BaseElementProps> {
     public static type: string;
