+++ packages/lib/src/components/types.ts
@@ -1,6 +1,6 @@
 import { Amount, AmountExtended } from '../types/shared';
 import Language from '../language/Language';
-import UIElement from './internal/UIElement/UIElement';
+import UIElement from './external/UIElement';
 import Core from '../core';
 import Analytics from '../core/Analytics';
 import BPSession from '../core/FPSession';
